requirements.txt
causalchamber==0.1.1
corner==2.2.3
matplotlib==3.8.4
nflows==0.14
numpy==2.2.6
pandas==2.2.3
Pillow==11.2.1
plotly==5.24.1
POT==0.9.5
PyYAML==6.0.1
PyYAML==6.0.2
PyYAML==6.0.2
scikit_learn==1.4.2
torch==2.3.1
torch==2.2.1
torchvision==0.18.1
torchvision==0.17.1
